Reorganized context tests to mirror folder structure in src folder.
