+++ deep_causality/tests/types/context_types/mod.rs
@@ -1,19 +1,13 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) "2023" . The DeepCausality Authors. All Rights Reserved.
 
+mod adjustable;
+mod context_graph;
+mod contextoid;
 mod node_types;
 
 #[cfg(test)]
-mod context_tests;
-#[cfg(test)]
-mod contextoid_tests;
-#[cfg(test)]
 mod time_scale_tests;
 #[cfg(test)]
-mod graph_node_tests;
-#[cfg(test)]
-mod graph_node_type_tests;
-#[cfg(test)]
-mod graph_relation_kind_tests;
-#[cfg(test)]
-mod graph_root_tests;
+mod relation_kind_tests;
+
